requirements.txt
pymongo
python-dotenv
dnspython
numpy
scikit-learn
matplotlib
